% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_column.R
\name{summarize_column}
\alias{summarize_column}
\title{Summarize columns of data}
\usage{
summarize_column(data_frame, column)
}
\arguments{
\item{data_frame}{A data frame or data frame extension}

\item{column}{unnamed column preparing to summarize}
}
\value{
A data frame with 3 columns (mean, med, sd). Given
if input is an empty df, then return an empty df
}
\description{
Create a new data frame with 3 columns (mean, med, sd)
}
\examples{
summarize_column(mtcars, am, mpg)
}
